web: server: Respond to a static file request with a bytevector.
If the response to a static file request is returned as a
procedure, (web server) tries to interpret the response as text. This
results in encoding errors with binary files. Fixing this is not worth
the trouble since (tissue web server) is anyway not meant to serve
large static files. Large static files should be served using a more
capable server such as nginx.

* tissue/web/server.scm (handler): Return response to a static file
request as a bytevector, instead of as a procedure.